+++ src/main/java/net/steveperkins/fitnessjiffy/service/ExerciseService.java
@@ -82,7 +82,7 @@ public final class ExerciseService {
                             weight.getPounds()
                     );
                     dto.setCaloriesBurned(caloriesBurned);
-                    final int pointsBurned = calculatePointsBurned(
+                    final double pointsBurned = calculatePointsBurned(
                             exercisePerformed.getExercise().getMetabolicEquivalent(),
                             exercisePerformed.getMinutes(),
                             weight.getPounds()
@@ -183,19 +183,19 @@ public final class ExerciseService {
         return (int) (metabolicEquivalent * 3.5 * weightInKilograms / 200 * minutes);
     }
 
-    private int calculatePointsBurned(
+    private double calculatePointsBurned(
             final double metabolicEquivalent,
             final int minutes,
             final double weightInPounds
     ) {
         final double caloriesBurnedPerHour = calculateCaloriesBurned(metabolicEquivalent, 60, weightInPounds);
-        int pointsBurned;
+        double pointsBurned;
         if (caloriesBurnedPerHour < 400) {
-            pointsBurned = (int) (weightInPounds * minutes * 0.000232);
+            pointsBurned = weightInPounds * minutes * 0.000232;
         } else if (caloriesBurnedPerHour < 900) {
-            pointsBurned = (int) (weightInPounds * minutes * 0.000327);
+            pointsBurned = weightInPounds * minutes * 0.000327;
         } else {
-            pointsBurned = (int) (weightInPounds * minutes * 0.0008077);
+            pointsBurned = weightInPounds * minutes * 0.0008077;
         }
         return pointsBurned;
     }
